Correct installation instructions
The previous installation instructions recommended installing the library to the Arduino IDE installation folder. This is a terrible idea because all libraries installed there are lost whenever you update to a new version of the Arduino IDE. The correct place to install libraries is to the libraries subfolder of the sketchbook folder.